refactor(output): extract formatting tests to sibling file
